// web/src/shared/api/generated/model/auditLogResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AuditLogResult {
  id: string;
  timestamp: string;
  /** @nullable */
  actorUserId: string | null;
  /** @nullable */
  actorUserName: string | null;
  /** @nullable */
  actorRole: string | null;
  action: string;
  resourceType: string;
  /** @nullable */
  resourceId: string | null;
  outcome: string;
  /** @nullable */
  errorCode: string | null;
  /** @nullable */
  traceId: string | null;
  /** @nullable */
  diff: string | null;
}